+++ src/lisp/kernel/cleavir/bir-to-bmir.lisp
@@ -220,9 +220,29 @@
 (deftransform core:two-arg->  core::two-arg-fixnum->  fixnum fixnum)
 (deftransform core:two-arg->= core::two-arg-fixnum->= fixnum fixnum)
 
+(deftransform truncate core::sf-truncate
+  (single-float #.(float most-negative-fixnum 1f0)
+                #.(float most-positive-fixnum 1f0))
+  single-float)
+(deftransform truncate core::df-truncate
+  (double-float #.(float most-negative-fixnum 1d0)
+                #.(float most-positive-fixnum 1d0))
+  double-float)
+(deftransform floor core::sf-truncate
+  (single-float 0f0 #.(float most-positive-fixnum 1f0))
+  single-float)
+(deftransform floor core::df-truncate
+  (double-float 0d0 #.(float most-positive-fixnum 1d0))
+  double-float)
+
 (deftransform ftruncate core::sf-ftruncate single-float single-float)
 (deftransform ftruncate core::df-ftruncate double-float double-float)
-;; TODO: One-arg form
+(deftransform ffloor core::sf-ffloor single-float single-float)
+(deftransform ffloor core::df-ffloor double-float double-float)
+(deftransform fceiling core::sf-fceil single-float single-float)
+(deftransform fceiling core::df-fceil double-float double-float)
+(deftransform fround core::sf-froundeven single-float single-float)
+(deftransform fround core::df-froundeven double-float double-float)
 
 (macrolet ((define-floatf (name sf-primop df-primop)
              `(progn
@@ -258,10 +278,13 @@
 ;;; When both dividend and divisor are positive, MOD and REM coincide,
 ;;; as do FLOOR and TRUNCATE.
 (deftransform truncate core::fixnum-truncate
-  fixnum (integer 1 #.most-positive-fixnum))
-(deftransform truncate core::fixnum-truncate
   ;; -2 because most-negative-fixnum/-1 would overflow
-  fixnum (integer #.most-negative-fixnum -2))
+  fixnum (or (integer #.most-negative-fixnum -2)
+           (integer 1 #.most-positive-fixnum)))
+;; division by -1 is ok as long as most-negative-fixnum is excluded
+(deftransform truncate core::fixnum-truncate
+  (integer #.(1+ most-negative-fixnum) #.most-positive-fixnum)
+  (and fixnum (not (eql 0))))
 (deftransform floor core::fixnum-truncate
   (integer 0 #.most-positive-fixnum) (integer 1 #.most-positive-fixnum))
 (deftransform mod core::fixnum-rem
@@ -313,20 +336,25 @@
   t fixnum t)
 ;; These are unsafe - make sure we only use core:vref when we don't need a
 ;; (further) bounds check.
-(defmacro define-vector-transforms (element-type)
-  `(progn
-     (deftransform core:vref (core:vref ,element-type)
-       (simple-array ,element-type (*)) fixnum)
-     (deftransform (setf core:vref) (core::vset ,element-type)
-       ;; FIXME: we should probably check the new value's type?
-       ;; ditto for atomic aref below.
-       t (simple-array ,element-type (*)) fixnum)))
-(define-vector-transforms t)
-(define-vector-transforms single-float)
-(define-vector-transforms double-float)
-(define-vector-transforms base-char)
-(define-vector-transforms character)
-(define-vector-transforms bit)
+(macrolet ((define-vector-transforms (element-type)
+             `(progn
+                (deftransform core:vref (core:vref ,element-type)
+                  (simple-array ,element-type (*)) fixnum)
+                (deftransform (setf core:vref) (core::vset ,element-type)
+                  ;; FIXME: we should probably check the new value's type?
+                  ;; ditto for atomic aref below.
+                  t (simple-array ,element-type (*)) fixnum))))
+  (define-vector-transforms t)
+  (define-vector-transforms single-float)
+  (define-vector-transforms double-float)
+  (define-vector-transforms base-char)
+  (define-vector-transforms character)
+  (define-vector-transforms ext:byte64) (define-vector-transforms ext:integer64)
+  (define-vector-transforms ext:byte32) (define-vector-transforms ext:integer32)
+  (define-vector-transforms ext:byte16) (define-vector-transforms ext:integer16)
+  (define-vector-transforms ext:byte8) (define-vector-transforms ext:integer8)
+  (define-vector-transforms fixnum)
+  (define-vector-transforms bit))
 
 (deftransform array-total-size core::vector-length (simple-array * (*)))
 
@@ -370,6 +398,7 @@
 (defun atomic-aset-test (return-type new-type order-type array-type
                          &rest index-types)
   (declare (ignore return-type new-type index-types))
+  ;; FIXME: check new-type
   (atomic-aref-test* array-type order-type
                      clasp-cleavir:*clasp-system*))