+++ src/lisp/kernel/cleavir/ir.lisp
@@ -99,6 +99,8 @@
 (defun %fptrunc (value type &optional (label ""))
   (llvm-sys:create-fptrunc cmp:*irbuilder* value type label))
 
+(defun %fptosi (value type &optional (label ""))
+  (llvm-sys:create-fpto-si cmp:*irbuilder* value type label))
 (defun %sitofp (value type &optional (label ""))
   (llvm-sys:create-sito-fp cmp:*irbuilder* value type label))